… chain/types/cbor_gen.go
… go.sum
… node/hello/cbor_gen.go
… paychmgr/cbor_gen.go