Worth looking at the cbor_gen.go files a little bit, but this improves the situations pretty significantly:

Current testnet/3:
```
why@whynix ~/c/lotus (testnet/3)> go test ./chain/types -bench=. -run=XXX
goos: linux
goarch: amd64
pkg: github.com/filecoin-project/lotus/chain/types
BenchmarkBlockHeaderMarshal-12    	  413956	      2601 ns/op	     880 B/op	      46 allocs/op
BenchmarkSerializeMessage-12      	  801396	      1364 ns/op	     640 B/op	      19 allocs/op
PASS
ok  	github.com/filecoin-project/lotus/chain/types	2.225s
```

This branch:
```
why@whynix ~/c/lotus (deps/update-cbor-gen)> go test ./chain/types -bench=. -run=XXX
goos: linux
goarch: amd64
pkg: github.com/filecoin-project/lotus/chain/types
BenchmarkBlockHeaderMarshal-12    	 1258426	       987 ns/op	      80 B/op	      12 allocs/op
BenchmarkSerializeMessage-12      	 1254820	       991 ns/op	     384 B/op	      12 allocs/op
PASS
ok  	github.com/filecoin-project/lotus/chain/types	4.349s
```